Made the server console app more interactive by adding a custom banner.
Added a clientConfig txt file to the resource folder. Users write the server ip address ( in IPV4 format!!!) they wish to connect to in clientConfig.txt. If an invalid string is provided or if the file is empty, the program defaults to localhost.
Hopefully removed the annoying go message after 3 seconds of showing up